Update to .NET 9

+++ AnnoDesigner/Models/AvailableRelease.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AnnoDesigner.Models
 {
     public class AvailableRelease
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public ReleaseType Type { get; set; }
